backend/requirements.txt
# Core Django framework
Django==5.2.7

# Django REST Framework for API development
djangorestframework==3.15.2

# JWT authentication for DRF
djangorestframework-simplejwt==5.3.0

# MySQL database connector (pure Python, Windows-friendly)
PyMySQL==1.1.2

# Required for pkg_resources (used by djangorestframework-simplejwt)
setuptools>=68.0.0

# Swagger/OpenAPI documentation
drf-spectacular==0.27.2

# CORS headers for frontend communication
django-cors-headers==4.4.0

# Core dependencies (automatically installed with Django)
asgiref==3.10.0
sqlparse==0.5.3
tzdata==2025.2

# Python JWT library (dependency of djangorestframework-simplejwt)
PyJWT==2.9.0

# Development dependencies (optional)
# Uncomment the following for development
# django-cors-headers==4.4.0  # For CORS handling in development
# python-decouple==3.8        # For environment variable management
# whitenoise==6.6.0           # For serving static files in production